refactor: 반환타입 long으로 수정 #22
