+++ src/main/java/server/acode/global/auth/AuthController.java
@@ -13,7 +13,7 @@ import server.acode.domain.user.repository.UserRepository;
 import server.acode.global.auth.dto.request.AccessTokenRequest;
 import server.acode.global.auth.dto.response.TokenResponse;
 import server.acode.global.auth.security.CustomUserDetails;
-import server.acode.global.auth.security.SecurityUtils;
+import server.acode.global.util.SecurityUtil;
 
 @RestController
 @RequiredArgsConstructor
@@ -36,7 +36,7 @@ public class AuthController {
     @GetMapping("/test/user/securityUtil")
     @Operation(summary = "유저 확인 테스트용", description = "개발자용입니다 토큰 넣고 호출 시 사용자 이름이 리턴됩니다")
     public void test(){
-        authService.checkUser(SecurityUtils.getCurrentUserId());
+        authService.checkUser(SecurityUtil.getCurrentUserId());
     }
 
     @Operation(summary = "로그아웃")
@@ -48,8 +48,7 @@ public class AuthController {
     @Operation(summary = "회원 탈퇴")
     @DeleteMapping("/withdrawal")
     public ResponseEntity withdrawal(HttpServletRequest request){
-        String userId = SecurityUtils.getCurrentUserId();
-        return authService.withdrawal(request.getHeader("Authorization"), userId);
+        return authService.withdrawal(request.getHeader("Authorization"), SecurityUtil.getCurrentUserId());
     }
 
     @Operation(summary = "access token 재발급")
